Fix wrongly restricted procedures.
