+++ core/src/test/java/apoc/ArgumentDescriptionsTest.java
@@ -18,7 +18,6 @@
  */
 package apoc;
 
-import static java.util.Collections.singletonList;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.Assert.fail;
 
@@ -66,9 +65,11 @@ import apoc.math.Maths;
 import apoc.math.Regression;
 import apoc.merge.Merge;
 import apoc.meta.Meta;
+import apoc.meta.MetaRestricted;
 import apoc.neighbors.Neighbors;
 import apoc.nodes.Grouping;
 import apoc.nodes.Nodes;
+import apoc.nodes.NodesRestricted;
 import apoc.number.ArabicRoman;
 import apoc.number.Numbers;
 import apoc.number.exact.Exact;
@@ -77,6 +78,7 @@ import apoc.path.Paths;
 import apoc.periodic.Periodic;
 import apoc.refactor.GraphRefactoring;
 import apoc.refactor.rename.Rename;
+import apoc.schema.SchemaRestricted;
 import apoc.schema.Schemas;
 import apoc.scoring.Scoring;
 import apoc.search.ParallelNodeSearch;
@@ -100,6 +102,7 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Comparator;
 import java.util.HashSet;
+import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import org.junit.After;
@@ -116,7 +119,25 @@ public class ArgumentDescriptionsTest {
 
     @Rule
     public DbmsRule db = new ImpermanentDbmsRule()
-            .withSetting(GraphDatabaseSettings.procedure_unrestricted, singletonList("apoc.*"));
+            .withSetting(
+                    GraphDatabaseSettings.procedure_unrestricted,
+                    List.of(
+                            "apoc.nodes.link",
+                            "apoc.node.relationship.exists",
+                            "apoc.nodes.connected",
+                            "apoc.nodes.isDense",
+                            "apoc.schema.nodes",
+                            "apoc.schema.relationship",
+                            "apoc.meta.nodes.count",
+                            "apoc.meta.stats",
+                            "apoc.meta.data",
+                            "apoc.meta.schema",
+                            "apoc.meta.nodeTypeProperties",
+                            "apoc.meta.relTypeProperties",
+                            "apoc.meta.graph",
+                            "apoc.meta.graph.of",
+                            "apoc.meta.graphSample",
+                            "apoc.meta.subGraph"));
 
     @Before
     public void setUp() {
@@ -164,9 +185,11 @@ public class ArgumentDescriptionsTest {
                 Median.class,
                 Merge.class,
                 Meta.class,
+                MetaRestricted.class,
                 Neighbors.class,
                 Neo4jLogStream.class,
                 Nodes.class,
+                NodesRestricted.class,
                 Numbers.class,
                 ParallelNodeSearch.class,
                 PathExplorer.class,
@@ -180,6 +203,7 @@ public class ArgumentDescriptionsTest {
                 Regression.class,
                 SchemaIndex.class,
                 Schemas.class,
+                SchemaRestricted.class,
                 Scoring.class,
                 Statistics.class,
                 Strings.class,
